+++ src/heap-profile-stats.h
@@ -58,10 +58,10 @@ struct HeapProfileStats {
         alloc_size - free_size == other.alloc_size - other.free_size;
   }
 
-  int32 allocs;      // Number of allocation calls.
-  int32 frees;       // Number of free calls.
-  int64 alloc_size;  // Total size of all allocated objects so far.
-  int64 free_size;   // Total size of all freed objects so far.
+  int64_t allocs;      // Number of allocation calls.
+  int64_t frees;       // Number of free calls.
+  int64_t alloc_size;  // Total size of all allocated objects so far.
+  int64_t free_size;   // Total size of all freed objects so far.
 };
 
 // Allocation and deallocation statistics per each stack trace.
